style: rename register action file and models -> schemas
